+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add buttons to open appropriate compendiums to the features, items and spells parts of actor sheets
 - Add Chat Popout functionality for attack and item chat cards
 - Attack chat cards now show the attack's icon
+- Attack and item chat cards now show the combat round they were shown, if in combat
 
 ## 0.75.7